Move GraknAdmin into different package
