+++ grakn-graph/src/main/java/ai/grakn/graph/internal/AbstractGraknGraph.java
@@ -19,7 +19,7 @@
 package ai.grakn.graph.internal;
 
 import ai.grakn.Grakn;
-import ai.grakn.GraknAdmin;
+import ai.grakn.graph.GraknAdmin;
 import ai.grakn.GraknGraph;
 import ai.grakn.concept.Concept;
 import ai.grakn.concept.ConceptId;
