+++ grakn-graph/src/main/java/ai/grakn/graph/internal/RuleTypeImpl.java
@@ -18,7 +18,7 @@
 
 package ai.grakn.graph.internal;
 
-import ai.grakn.GraknAdmin;
+import ai.grakn.graph.GraknAdmin;
 import ai.grakn.concept.Rule;
 import ai.grakn.concept.RuleType;
 import ai.grakn.graql.Pattern;
